Feature/endpoints

+++ ArticleServiceAPI/Model/ArticleDTO.cs
@@ -9,7 +9,7 @@ namespace ArticleServiceAPI.Model
     {
         public string? Name { get; set; }
         public bool NoReserve { get; set; }
-        public float EstimatedPrice { get; set; }
+        public double EstimatedPrice { get; set; }
         public string? Description { get; set; }
         public string? Category { get; set; }
         public bool Sold { get; set; }
